+++ CHANGELOG.md
@@ -12,3 +12,4 @@ Created on Dec 20, 2023
 - Initial Development: Development started on Decemeber 20, 2023 -> **0.1**
 - Initial Development: Authentication processes work on Decemeber 21, 2023 -> **0.2**
 - Initial Development: Account manager is completed on Decemeber 21, 2023 -> **0.22**
+- Initial Development: Implementing transaction system on Decemeber 22, 2023 -> **0.3**